refactor: NumericGeneartor 패키지 이동 (controller -> model)
