+++ src/main/java/racingcar/model/Cars.java
@@ -1,7 +1,5 @@
 package racingcar.model;
 
-import racingcar.controller.NumericGenerator;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Matcher;
