+++ src/main/java/racingcar/model/numericgenerator/RandomNumericGenerator.java
@@ -1,8 +1,8 @@
-package racingcar.controller.numericgenerator;
-
-import racingcar.controller.NumericGenerator;
+package racingcar.model.numericgenerator;
 
 import java.util.Random;
+
+import racingcar.model.NumericGenerator;
 
 public class RandomNumericGenerator implements NumericGenerator {
     private final Random random = new Random();
