+++ src/test/java/racingcar/model/CarsTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.controller.NumericGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
